Merge 31 commits from upstream

+++ PKHeX.Core/Legality/Encounters/Generator/Possible/EncounterPossible6.cs
@@ -232,7 +232,7 @@ public record struct EncounterPossible6(EvoCriteria[] Chain, EncounterTypeGroup 
         for (; Index < db.Length;)
         {
             var enc = db[Index++];
-            if (!enc.CanBeReceivedByVersion((int)Version))
+            if (!enc.CanBeReceivedByVersion(Version))
                 continue;
             foreach (var evo in Chain)
             {
