+++ PKHeX.Core/Legality/Encounters/Verifiers/EggHatchLocation8b.cs
@@ -13,14 +13,14 @@ public static class EggHatchLocation8b
     /// <summary>
     /// Returns true if the hatch location is valid for Brilliant Diamond.
     /// </summary>
-    public static bool IsValidMet8BD(int location) => HasLocationFlag(LocationPermitted8b, MaskBD, location);
+    public static bool IsValidMet8BD(ushort location) => HasLocationFlag(LocationPermitted8b, MaskBD, location);
 
     /// <summary>
     /// Returns true if the hatch location is valid for Shining Pearl.
     /// </summary>
-    public static bool IsValidMet8SP(int location) => HasLocationFlag(LocationPermitted8b, MaskSP, location);
+    public static bool IsValidMet8SP(ushort location) => HasLocationFlag(LocationPermitted8b, MaskSP, location);
 
-    private static bool HasLocationFlag(ReadOnlySpan<byte> arr, byte mask, int location)
+    private static bool HasLocationFlag(ReadOnlySpan<byte> arr, byte mask, ushort location)
     {
         if ((uint)location >= arr.Length)
             return false;
