+++ PKHeX.Core/PKM/Util/Conversion/FormConverter.cs
@@ -22,7 +22,7 @@ public static class FormConverter
     /// <returns>A list of strings corresponding to the forms that a Pokémon can have.</returns>
     public static string[] GetFormList(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, IReadOnlyList<string> genders, EntityContext context)
     {
-        int generation = context.Generation();
+        byte generation = context.Generation();
 
         // Mega List
         if (context.IsMegaGeneration() && IsFormListSingleMega(species))
@@ -52,7 +52,7 @@ public static class FormConverter
 
     private static string[] GetFormsGen1(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, EntityContext context)
     {
-        int generation = context.Generation();
+        byte generation = context.Generation();
         return (Species)species switch
         {
             Charizard or Mewtwo when context.IsMegaGeneration() => GetMegaXY(types, forms),
@@ -74,7 +74,7 @@ public static class FormConverter
 
     private static string[] GetFormsGen2(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, EntityContext context)
     {
-        int generation = context.Generation();
+        byte generation = context.Generation();
         return (Species)species switch
         {
             Pichu when context is Gen4 => GetFormsPichu(types, forms),
@@ -86,7 +86,7 @@ public static class FormConverter
         };
     }
 
-    private static string[] GetFormsGen3(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, int generation)
+    private static string[] GetFormsGen3(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, byte generation)
     {
         return (Species)species switch
         {
@@ -111,7 +111,7 @@ public static class FormConverter
         };
     }
 
-    private static string[] GetFormsGen4(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, int generation)
+    private static string[] GetFormsGen4(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, byte generation)
     {
         return (Species)species switch
         {
@@ -153,7 +153,7 @@ public static class FormConverter
         };
     }
 
-    private static string[] GetFormsGen5(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, int generation)
+    private static string[] GetFormsGen5(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, byte generation)
     {
         return (Species)species switch
         {
@@ -212,7 +212,7 @@ public static class FormConverter
         };
     }
 
-    private static string[] GetFormsGen6(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, IReadOnlyList<string> genders, int generation)
+    private static string[] GetFormsGen6(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, IReadOnlyList<string> genders, byte generation)
     {
         return (Species)species switch
         {
@@ -308,7 +308,7 @@ public static class FormConverter
         };
     }
 
-    private static string[] GetFormsGen7(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, int generation)
+    private static string[] GetFormsGen7(ushort species, IReadOnlyList<string> types, IReadOnlyList<string> forms, byte generation)
     {
         return (Species)species switch
         {
@@ -372,7 +372,7 @@ public static class FormConverter
         };
     }
 
-    private static string[] GetFormsGen8(ushort species, int generation, IReadOnlyList<string> types, IReadOnlyList<string> forms, IReadOnlyList<string> genders)
+    private static string[] GetFormsGen8(ushort species, byte generation, IReadOnlyList<string> types, IReadOnlyList<string> forms, IReadOnlyList<string> genders)
     {
         return (Species)species switch
         {
@@ -449,7 +449,7 @@ public static class FormConverter
         };
     }
 
-    private static string[] GetFormsGen9(ushort species, int generation, IReadOnlyList<string> types, IReadOnlyList<string> forms, IReadOnlyList<string> genders)
+    private static string[] GetFormsGen9(ushort species, byte generation, IReadOnlyList<string> types, IReadOnlyList<string> forms, IReadOnlyList<string> genders)
     {
         return (Species)species switch
         {
@@ -527,7 +527,7 @@ public static class FormConverter
 
     private static string[] GetFormsAlolan(EntityContext context, IReadOnlyList<string> types, IReadOnlyList<string> forms, ushort species)
     {
-        int generation = context.Generation();
+        byte generation = context.Generation();
         if (generation < 7)
             return EMPTY;
 
@@ -559,7 +559,7 @@ public static class FormConverter
 
     private static string[] GetFormsPikachu(EntityContext context, IReadOnlyList<string> types, IReadOnlyList<string> forms)
     {
-        int generation = context.Generation();
+        byte generation = context.Generation();
         return generation switch
         {
             6 => [
@@ -617,7 +617,7 @@ public static class FormConverter
         ];
     }
 
-    private static string[] GetFormsArceus(ushort species, int generation, IReadOnlyList<string> types, IReadOnlyList<string> forms)
+    private static string[] GetFormsArceus(ushort species, byte generation, IReadOnlyList<string> types, IReadOnlyList<string> forms)
     {
         return generation switch
         {
@@ -727,7 +727,7 @@ public static class FormConverter
         ],
     };
 
-    private static string[] GetFormsUnown(int generation) => generation switch
+    private static string[] GetFormsUnown(byte generation) => generation switch
     {
         2 =>
         [
@@ -790,7 +790,7 @@ public static class FormConverter
         ];
     }
 
-    private static string[] GetFormsHisui(ushort species, int generation, IReadOnlyList<string> types, IReadOnlyList<string> forms) => generation switch
+    private static string[] GetFormsHisui(ushort species, byte generation, IReadOnlyList<string> types, IReadOnlyList<string> forms) => generation switch
     {
         8 => (Species)species switch
         {
